% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABF1_sacCer3}
\alias{ABF1_sacCer3}
\title{ABF1_sacCer3}
\format{
An object of class \code{"GRanges"}.
}
\usage{
data(ABF1_sacCer3)
}
\description{
Genomic loci with a REB1 binding motifs according to 
http://jaspar.genereg.net/api/v1/matrix/MA0265.1.jaspar. PWM and scanning
done with TFBSTools.
}
\examples{
data(ABF1_sacCer3)
ABF1_sacCer3
}
\references{
Rossi, Lai & Pugh 2018 Genome Research
}
\keyword{datasets}
