% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATAC_ce11_Serizay2020}
\alias{ATAC_ce11_Serizay2020}
\title{ATAC_ce11_Serizay2020}
\format{
An object of class \code{"list"}.
}
\usage{
data(ATAC_ce11_Serizay2020)
}
\description{
A sample of ATAC-seq fragments from individual worm tissues (Serizay et 
al. 2020, "Tissue-specific profiling reveals distinctive regulatory 
architectures for ubiquitous, germline and somatic genes", BiorXiv)
}
\examples{
data(ATAC_ce11_Serizay2020)
ATAC_ce11_Serizay2020
}
\keyword{datasets}
