% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bam_test}
\alias{bam_test}
\title{bam_test}
\format{
An object of class \code{"GRanges"}.
}
\usage{
data(bam_test)
}
\description{
A .bam file sample
}
\examples{
data(bam_test)
bam_test
}
\keyword{datasets}
