% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat.R
\name{computeVmat}
\alias{computeVmat}
\title{A function to compute Vmat.}
\usage{
computeVmat(reads.granges, target.granges, XAXIS.CUTOFF = 1000,
  YAXIS.CUTOFF = 1000, stranded = TRUE)
}
\arguments{
\item{reads.granges}{GRanges. The paired-end fragments}

\item{target.granges}{GRanges. The regions to map the fragments onto}

\item{XAXIS.CUTOFF}{single integer. The x limits (-/+) of the computed Vmat}

\item{YAXIS.CUTOFF}{single integer. The y upper limit of the computed Vmat}

\item{stranded}{Boolean. Should the Vmat be stranded? If TRUE, negative-strand 
features are flipped.}
}
\value{
A table object
}
\description{
A function to compute Vmat.
}
