% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTCF_hg38}
\alias{CTCF_hg38}
\title{CTCF_hg38}
\format{
An object of class \code{"GRanges"}.
}
\usage{
data(CTCF_hg38)
}
\description{
high-score CTCF binding motifs, obtained from JASPAR
}
\examples{
data(CTCF_hg38)
CTCF_hg38
}
\keyword{datasets}
