% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deconvolveBidirectionalPromoters}
\alias{deconvolveBidirectionalPromoters}
\title{A function to duplicate unstranded GRanges into '+' and '-' GRanges}
\usage{
deconvolveBidirectionalPromoters(granges)
}
\arguments{
\item{granges}{A GRanges object.}

\item{names}{A character vector, length(names) == ncol(df).}
}
\value{
A data frame with new column names
}
\description{
A function to duplicate unstranded GRanges into '+' and '-' GRanges
}
