% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragsize.R
\name{get.distr}
\alias{get.distr}
\title{A fragsize function to compute fragments sizes}
\usage{
get.distr(bam.list, granges, extend.granges = c(-2000, 2000),
  limits = c(50, 650))
}
\arguments{
\item{bam.list}{Output of import.bam function.}

\item{granges}{A GRanges object. The genomic ranges to which the fragments 
should overlap in order to be considered when estimating fragment sizes.}

\item{extend.granges}{A numeric vector of length 2. How the granges should
be extended.}

\item{limits}{A numeric vector of length 2. Only consider fragments within
this window of sizes.}
}
\value{
A list of tbl, one for each .bam file.
}
\description{
\code{get.distr(bam.list)} returns the number of fragments for each fragment 
  length between \code{limits[1]} and \code{limits[2]} bp.
}
\examples{
read.bam(file, remove.duplicates = T, paired.reads = F filter.lengths = 
  c(50, 250))

}
