% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromSizes.default}
\alias{getChromSizes.default}
\title{A function to get the size of chromosomes for the input genome}
\usage{
\method{getChromSizes}{default}(genome = c("ce11", "dm6", "mm10", "hg38",
  "sacCer3", "danRer10"))
}
\arguments{
\item{genome}{a BSgenome object}
}
\value{
GRanges of whole chromosomes for the input genome
}
\description{
A function to get the size of chromosomes for the input genome
}
