% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromSizes.GRanges}
\alias{getChromSizes.GRanges}
\title{A function to estimate the size of chromosomes for the input genome}
\usage{
\method{getChromSizes}{GRanges}(granges)
}
\arguments{
\item{genome}{a GRanges object}
}
\value{
GRanges of whole chromosomes for the input genome
}
\description{
A function to estimate the size of chromosomes for the input genome
}
