% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromSizes}
\alias{getChromSizes}
\title{A function to get / estimate the size of chromosomes for the input genome}
\usage{
getChromSizes(x, ...)
}
\value{
GRanges of whole chromosomes for the input genome
}
\description{
A function to get / estimate the size of chromosomes for the input genome
}
