% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint.R
\name{getCuts}
\alias{getCuts}
\title{Internal function}
\usage{
getCuts(gr)
}
\arguments{
\item{gr}{GRanges Paired-end fragments used to extract their extremities}
}
\value{
GRanges
}
\description{
Function to extract cuts (i.e. extremities) of fragments stored
as GRanges.
}
