% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamutils.R
\name{importBam}
\alias{importBam}
\title{A bamutils function to import a single .bam file}
\usage{
importBam(files, remove.duplicates = T, paired.reads = F,
  filter.lengths = c(0, 0), quality.cutoff = 10L)
}
\arguments{
\item{remove.duplicates}{A boolean value. Should the duplicate reads 
(mapping to the same coordinates) be removed?}

\item{paired.reads}{A boolean value. Are the reads paired-end?}

\item{filter.lengths}{A numeric vector of length 2. The two values indicate
the minimum / maximum size of the reads to be returned.}

\item{quality.cutoff}{An integer. The minimum mapping quality of the reads
to be returned.}

\item{file}{A character vector. Each element is the path of an individual
.bam file.}
}
\value{
A GRanges object containing reads from the .bam file.
}
\description{
\code{read.bam(file)} returns the reads contained in the .bam file in a 
  GRanges object. Tested .bam files have been generated by BWA and follow
  the standard BAM format.
}
\examples{
read.bam(file, remove.duplicates = T, paired.reads = F filter.lengths = 
  c(50, 250))

}
