% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MNase_sacCer3_Henikoff2011}
\alias{MNase_sacCer3_Henikoff2011}
\title{MNase_sacCer3_Henikoff2011}
\format{
An object of class \code{"GRanges"}.
}
\usage{
data(MNase_sacCer3_Henikoff2011)
}
\description{
A sample of MNase-seq fragments from yeast (Henikoff et 
al. 2011, "Epigenome characterization at single base-pair resolution", 
PNAS)
}
\examples{
data(MNase_sacCer3_Henikoff2011)
MNase_sacCer3_Henikoff2011
}
\keyword{datasets}
