% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MNase_sacCer3_Henikoff2011_subset}
\alias{MNase_sacCer3_Henikoff2011_subset}
\title{MNase_sacCer3_Henikoff2011_subset}
\format{
An object of class \code{"GRanges"}.
}
\usage{
data(MNase_sacCer3_Henikoff2011_subset)
}
\description{
A sample of fragments from multiple MNase-seq experiments performed in 
yeast (Henikoff et al. 2011, "Epigenome characterization at single 
base-pair resolution", PNAS), mapping over chrXV:186,400-187,400.
}
\examples{
data(MNase_sacCer3_Henikoff2011_subset)
MNase_sacCer3_Henikoff2011_subset
}
\keyword{datasets}
