% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{namedListToLongFormat}
\alias{namedListToLongFormat}
\title{A function to easily coerce a named list into a long data.frame}
\usage{
namedListToLongFormat(x)
}
\arguments{
\item{x}{A named list vector.}
}
\value{
A long data frame
}
\description{
\code{namedListToLongFormat(x)} returns a data.frame in long format, with 
an added 'name' column, containing the names of the input list.
}
