% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat.R
\name{normalizeVmat}
\alias{normalizeVmat}
\title{A function to normalized a Vmat to a given background.}
\usage{
normalizeVmat(Vmat, bam_granges, normFun = c("libdepth", "sum", "max",
  "quantile", "Zscore"), universe = NULL, libdepth = NULL, s = 0.999,
  roll = 1)
}
\arguments{
\item{Vmat}{A Vmat, usually output of computeVmat}

\item{normFun}{string A Vmat can be scaled relative to its median ('pctmedian'), 
to its mean ('pctmean'), to its max ('pctmax'). Otherwise it could be zscore-d
entirely ('Zscore') or by rows ('rowZscore') or by columns ('colZscore').}

\item{roll}{integer to use as the window to smooth the Vmat rows by rolling mean}
}
\value{
A normalized Vmat object
}
\description{
A function to normalized a Vmat to a given background.
}
