% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat.R
\name{nucleosomeEnrichment.default}
\alias{nucleosomeEnrichment.default}
\title{A function to compute nucleosome enrichment over a set of GRanges}
\usage{
\method{nucleosomeEnrichment}{default}(Vmat, background, ...)
}
\arguments{
\item{Vmat}{a computed Vmat. Does not need to be normalized.}

\item{background}{a background Vmat}
}
\value{
list
}
\description{
A function to compute nucleosome enrichment over a set of GRanges
}
