% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosome_enrichment.R
\name{nucleosomeEnrichment}
\alias{nucleosomeEnrichment}
\title{A function to compute nucleosome enrichment over a set of GRanges}
\usage{
nucleosomeEnrichment(x, ...)
}
\arguments{
\item{x}{a GRanges or Vmat}

\item{...}{additional parameters}
}
\value{
list
}
\description{
A function to compute nucleosome enrichment over a set of GRanges
}
\examples{
data(bam_test)
data(ce11_proms)
nucleosomeEnrichment(bam_test, ce11_proms)
}
