% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragsize.R
\name{plotFragmentsDistribution}
\alias{plotFragmentsDistribution}
\title{A fragsize function to plot fragments sizes}
\usage{
plotFragmentsDistribution(sizes)
}
\arguments{
\item{distr.list}{Output of get.distr. A list of tbl, one for each .bam file.}
}
\value{
A list of ggplots
}
\description{
A fragsize function to plot fragments sizes
}
