% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat.R
\name{plotVmat.Vmat}
\alias{plotVmat.Vmat}
\title{A function to plot a computed Vmat}
\usage{
\method{plotVmat}{Vmat}(x, ...)
}
\arguments{
\item{x}{A computed Vmat (ideally, should be normalized)}

\item{...}{additional parameters}
}
\value{
A Vmat ggplot
}
\description{
The plotVmat.Vmat() method forwards the Vmat to plotVmat.default().
}
\examples{
data(bam_test)
data(ce11_proms)
V <- plotVmat(
    bam_test,
    ce11_proms,
    normFun = 'libdepth+nloci', 
    return_Vmat = TRUE
)
plotVmat(V)
}
