% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REB1_sacCer3}
\alias{REB1_sacCer3}
\title{REB1_sacCer3}
\format{
An object of class \code{"GRanges"}.
}
\usage{
data(REB1_sacCer3)
}
\description{
High-confidence REB1 binding sites, from 
Rossi, Lai & Pugh 2018 Genome Research
}
\examples{
data(REB1_sacCer3)
REB1_sacCer3
}
\references{
Rossi, Lai & Pugh 2018 Genome Research
}
\keyword{datasets}
