% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat.R
\name{scaleVmat}
\alias{scaleVmat}
\title{A function to scale a Vmat.}
\usage{
scaleVmat(Vmat, FUN = "pctsum")
}
\arguments{
\item{Vmat}{}

\item{FUN}{string. A Vmat can be scaled relative to its median ('pctmedian'), 
to its mean ('pctmean'), to its max ('pctmax'). Otherwise it could be zscore-d
entirely ('Zscore') or by rows ('rowZscore') or by columns ('colZscore').}
}
\value{
A table object
}
\description{
A function to scale a Vmat.
}
