% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqinfos}
\alias{seqinfos}
\title{seqinfos}
\format{
An object of class \code{"Seqinfo"}.
}
\usage{
data(seqinfos)
}
\description{
Chrom sizes fom different genomes
}
\examples{
data(seqinfos)
seqinfos
}
\keyword{datasets}
