% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setColNames}
\alias{setColNames}
\title{A function to easily set column names for a data.frame}
\usage{
setColNames(df, names)
}
\arguments{
\item{df}{A data frame.}

\item{names}{A character vector, length(names) == ncol(df).}
}
\value{
A data frame with new column names
}
\description{
\code{setColNames(df, names)} returns a data.frame with new column names.
This function can be convinently used in the tidyverse.
}
