% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffleGRanges}
\alias{shuffleGRanges}
\title{A function to shuffle GRanges along a given genome
Internaly, this function relies on bedTools shuffle, so the bedTools
suite has to be installed on the workstation.}
\usage{
shuffleGRanges(granges, genome = NULL,
  opt.shuffle = "-chrom -noOverlapping", exclude_itself = TRUE,
  SEED = 222)
}
\arguments{
\item{granges}{A GRanges object to shuffle}

\item{genome}{a BSgenome object. See getChromSizes for more details}

\item{opt.shuffle}{string of options (in single quotes) to pass to bedtools shuffle}

\item{exclude_itself}{Boolean. Should the shuffled GRanges overlap the input GRanges or not?}

\item{SEED}{Integer. Make shuffling reproducible}
}
\value{
a GRanges object with shuffled ranges
}
\description{
A function to shuffle GRanges along a given genome
Internaly, this function relies on bedTools shuffle, so the bedTools
suite has to be installed on the workstation.
}
