% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat_utils.R
\name{shuffleVmat}
\alias{shuffleVmat}
\title{A function to shuffle a Vmat}
\usage{
shuffleVmat(Vmat)
}
\arguments{
\item{Vmat}{A Vmat, usually output of computeVmat}
}
\value{
A shuffled Vmat object
}
\description{
This function works on a Vmat (the output of computeVmat()). It shuffles
the matrix to randomize the fragment densities.
}
\examples{
data(bam_test)
data(ce11_all_REs)
Vmat <- computeVmat(bam_test, ce11_all_REs)
Vmat <- shuffleVmat(Vmat)
}
