% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplifyGRanges}
\alias{simplifyGRanges}
\title{A function to remove metadata columns}
\usage{
simplifyGRanges(granges)
}
\arguments{
\item{granges}{A GRanges object}
}
\value{
GRanges without any metadata columns
}
\description{
A function to remove metadata columns
}
