% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alignToTSS}
\alias{alignToTSS}
\title{A function to re-align GRanges to their TSS}
\usage{
alignToTSS(granges, upstream, downstream)
}
\arguments{
\item{granges}{A GRanges object with a TSS column or TSS.rev and TSS.fwd columns}

\item{upstream}{How many bases upstream of the TSS?}

\item{downstream}{How many bases downstream of the TSS?}
}
\value{
GRanges aligned to the TSS column or to TSS.rev and TSS.fwd columns
}
\description{
A function to re-align GRanges to their TSS
}
