% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragsize.R
\name{getFragmentsDistribution}
\alias{getFragmentsDistribution}
\title{A function to compute sizes distribution for PE fragments}
\usage{
getFragmentsDistribution(fragments, granges_list = NULL,
  extend_granges = c(-500, 500), limits = c(0, 600))
}
\arguments{
\item{fragments}{GRanges object containing PE fragments. See importPEBamFiles
for more details.}

\item{extend_granges}{A numeric vector of length 2. How the GRanges should
be extended.}

\item{limits}{A numeric vector of length 2. Only consider fragments within
this window of sizes.}

\item{granges}{GRanges. cam be a list of different sets of GRanges.
should overlap in order to be considered when estimating fragment sizes.}
}
\value{
A list of tbl, one for each .bam file.
}
\description{
A function to compute sizes distribution for PE fragments
}
