% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat.R
\name{nucleosomeEnrichment.GRanges}
\alias{nucleosomeEnrichment.GRanges}
\title{A function to compute nucleosome enrichment over a set of GRanges}
\usage{
\method{nucleosomeEnrichment}{GRanges}(bam_granges, granges,
  estimate_background = TRUE, plus1_nuc_only = FALSE, verbose = TRUE,
  genome = "ce11", ...)
}
\arguments{
\item{bam_granges}{GRanges. The paired-end fragments}

\item{granges}{GRanges. The regions to map the fragments onto}

\item{estimate_background}{Boolean Should the background be estimated automatically?}

\item{plus1_nuc_only}{Should compute nucleosome enrichment only for +1 
nucleosome?}

\item{verbose}{Boolean}

\item{genome}{a BSgenome object. See getChromSizes for more details}
}
\value{
list
}
\description{
A function to compute nucleosome enrichment over a set of GRanges
}
