% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat.R
\name{nucleosomeEnrichment}
\alias{nucleosomeEnrichment}
\title{A function to compute nucleosome enrichment over a set of GRanges}
\usage{
nucleosomeEnrichment(x, ...)
}
\value{
list
}
\description{
A function to compute nucleosome enrichment over a set of GRanges
}
