% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat.R
\name{plotVmat.GRanges}
\alias{plotVmat.GRanges}
\title{A function to plot a computed Vmat}
\usage{
\method{plotVmat}{GRanges}(bam_granges, granges, XAXIS.CUTOFF = 1000,
  YAXIS.CUTOFF = 1000, stranded = TRUE, normalize = TRUE,
  normFun = "libdepth", universe = NULL, libdepth = NULL,
  s = 0.999, roll = 3, return_Vmat = FALSE, genome = "ce11", ...)
}
\arguments{
\item{bam_granges}{GRanges. The paired-end fragments}

\item{granges}{GRanges. The regions to map the fragments onto}

\item{XAXIS.CUTOFF}{single integer. The x limits (-/+) of the computed Vmat}

\item{YAXIS.CUTOFF}{single integer. The y upper limit of the computed Vmat}

\item{stranded}{Boolean. Should the Vmat be stranded? If TRUE, negative-strand 
features are flipped}

\item{normalize}{Boolean Should the Vmat be normalized?}

\item{normFun}{string A Vmat can be scaled relative to its median ('pctmedian'), 
to its mean ('pctmean'), to its max ('pctmax'). Otherwise it could be zscore-d
entirely ('Zscore') or by rows ('rowZscore') or by columns ('colZscore').}

\item{roll}{integer to use as the window to smooth the Vmat rows by rolling mean}

\item{return_Vmat}{Boolean. Should the function return the computed Vmat 
rather than the plot?}

\item{genome}{a BSgenome object. See getChromSizes for more details}

\item{Vmat2}{table A Vmat pre-computed over a background}

\item{estimate_background}{Boolean Should the background be estimated automatically?}

\item{background_granges}{GRanges The genomic loci used to compute the background Vmat}
}
\value{
A Vmat ggplot
}
\description{
A function to plot a computed Vmat
}
