% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmat.R
\name{shuffleVmat}
\alias{shuffleVmat}
\title{A function to shuffle a Vmat}
\usage{
shuffleVmat(Vmat, SEED = 999)
}
\arguments{
\item{Vmat}{A Vmat, usually output of computeVmat}

\item{background}{A Vmat computed from a background.}

\item{scale}{Boolean Should the background be removed?}

\item{normFun}{string A Vmat can be scaled relative to its median ('pctmedian'), 
to its mean ('pctmean'), to its max ('pctmax'). Otherwise it could be zscore-d
entirely ('Zscore') or by rows ('rowZscore') or by columns ('colZscore').}

\item{roll}{integer to use as the window to smooth the Vmat rows by rolling mean}
}
\value{
A normalized Vmat object
}
\description{
A function to shuffle a Vmat
}
